fix: move country detector to searchbar
