+++ libs/journeys/ui/src/components/SearchBar/SearchDropdown/SearchbarDropdown.tsx
@@ -48,6 +48,7 @@ interface SearchbarDropdownProps {
   open: boolean
   refinements: RefinementListRenderState
   id?: string
+  countryCode?: string
   anchorEl?: HTMLElement | null
   tabIndex?: number
   handleTabValueChange: Dispatch<SetStateAction<number>>
@@ -57,6 +58,7 @@ export function SearchbarDropdown({
   open,
   refinements,
   id,
+  countryCode,
   anchorEl,
   tabIndex: tabValue = 0,
   handleTabValueChange: setTabValue
@@ -119,7 +121,10 @@ export function SearchbarDropdown({
         sx={{ px: 8, pb: 8, bgcolor: 'background.paper', mt: 3 }}
         color="text.primary"
       >
-        <CountryLanguageSelector refinements={refinements} />
+        <CountryLanguageSelector
+          countryCode={countryCode}
+          refinements={refinements}
+        />
         <Box sx={{ borderBottom: 1, borderColor: 'divider', mb: 5 }}>
           <Tabs
             value={tabValue}
